Remove testUtils
