+++ packages/w3c-webdriver/test/10-command-contexts.test.ts
@@ -1,6 +1,20 @@
 import { browserName } from '../test-env/browser';
 import { session } from '../test-env/session';
-import { createWindow } from './testUtils';
+
+async function createWindow(): Promise<string> {
+  const handlesBefore = await session.getWindowHandles();
+  await session.executeScript(`window.open()`);
+  const handlesAfter = await session.getWindowHandles();
+  const newHandle = handlesAfter.find((handle: string) => !handlesBefore.includes(handle));
+
+  if (!newHandle) {
+    throw new Error('Creating new Window was not successful');
+  }
+
+  await session.switchToWindow(newHandle);
+
+  return newHandle;
+}
 
 describe('Command Contexts', () => {
   describe('getWindowHandle', () => {
